Refactor response classes
Introduce `JsonSchemaResponse` and consolidate response classes under
`aiida_restapi.common.responses`.
