+++ aiida_restapi/routers/computers.py
@@ -9,12 +9,12 @@ from aiida.cmdline.utils.decorators import with_dbenv
 from fastapi import APIRouter, Depends, Query, Request
 
 from aiida_restapi.common import query
+from aiida_restapi.common.responses import JsonApiResponse, JsonSchemaResponse
 from aiida_restapi.common.types import EntityIdentifier
 from aiida_restapi.jsonapi.adapters import JsonApiAdapter as JsonApi
 from aiida_restapi.jsonapi.models import errors
 from aiida_restapi.jsonapi.models.aiida import ComputerCollectionDocument, ComputerResourceDocument
 from aiida_restapi.jsonapi.models.base import JsonApiResourceDocument
-from aiida_restapi.jsonapi.responses import JsonApiResponse
 from aiida_restapi.services.entity import EntityService
 
 read_router = APIRouter(prefix='/computers')
@@ -25,6 +25,7 @@ service = EntityService[orm.Computer, orm.Computer.ReadModel](orm.Computer)
 
 @read_router.get(
     '/schema',
+    response_class=JsonSchemaResponse,
     response_model=dict[str, t.Any],
     responses={
         422: {'model': errors.RequestValidationError, 'description': 'Validation Error'},
