+++ aiida_restapi/routers/groups.py
@@ -9,6 +9,7 @@ from aiida.cmdline.utils.decorators import with_dbenv
 from fastapi import APIRouter, Depends, Query, Request
 
 from aiida_restapi.common import query
+from aiida_restapi.common.responses import JsonApiResponse, JsonSchemaResponse
 from aiida_restapi.common.types import EntityIdentifier
 from aiida_restapi.jsonapi.adapters import JsonApiAdapter as JsonApi
 from aiida_restapi.jsonapi.models import errors
@@ -19,7 +20,6 @@ from aiida_restapi.jsonapi.models.aiida import (
     UserResourceDocument,
 )
 from aiida_restapi.jsonapi.models.base import JsonApiResourceDocument
-from aiida_restapi.jsonapi.responses import JsonApiResponse
 from aiida_restapi.services.entity import EntityService
 
 read_router = APIRouter(prefix='/groups')
@@ -30,6 +30,7 @@ service = EntityService[orm.Group, orm.Group.ReadModel](orm.Group)
 
 @read_router.get(
     '/schema',
+    response_class=JsonSchemaResponse,
     response_model=dict[str, t.Any],
     responses={
         422: {'model': errors.RequestValidationError, 'description': 'Validation Error'},
