+++ aiida_restapi/routers/nodes.py
@@ -19,13 +19,13 @@ from typing_extensions import TypeAlias
 
 from aiida_restapi.common import exceptions as restapi_exceptions
 from aiida_restapi.common import query
+from aiida_restapi.common.responses import JsonApiResponse, JsonSchemaResponse
 from aiida_restapi.common.types import EntityIdentifier
 from aiida_restapi.config import API_CONFIG
 from aiida_restapi.jsonapi.adapters import JsonApiAdapter as JsonApi
 from aiida_restapi.jsonapi.errors import jsonapi_error
 from aiida_restapi.jsonapi.models import aiida, errors
 from aiida_restapi.jsonapi.models.base import JsonApiResourceDocument
-from aiida_restapi.jsonapi.responses import JsonApiResponse
 from aiida_restapi.models.node import NodeModelRegistry, NodeStatistics, NodeType
 from aiida_restapi.services.node import NodeService
 
@@ -75,6 +75,7 @@ def handle_request_validation_errors(
 
 @read_router.get(
     '/schema',
+    response_class=JsonSchemaResponse,
     response_model=dict[str, t.Any],
     responses={
         422: {
@@ -117,7 +118,10 @@ async def get_nodes_schema(
 async def get_node_projections(
     node_type: t.Annotated[
         str | None,
-        Query(description='The AiiDA node type string.', alias='type'),
+        Query(
+            description='The AiiDA node type string.',
+            alias='type',
+        ),
     ] = None,
 ) -> list[str]:
     """Get queryable projections for AiiDA nodes."""
